Fix compilation warnings on OSX

+++ lyaml.c
@@ -92,7 +92,7 @@ static void generate_error_message(struct lua_yaml_loader *loader) {
    luaL_addstring(&b, buf);
 
    if (loader->parser.problem_mark.line || loader->parser.problem_mark.column) {
-      snprintf(buf, sizeof(buf), ", line: %d, column: %d\n",
+      snprintf(buf, sizeof(buf), ", line: %zd, column: %zd\n",
          loader->parser.problem_mark.line + 1,
          loader->parser.problem_mark.column + 1);
       luaL_addstring(&b, buf);
@@ -101,7 +101,7 @@ static void generate_error_message(struct lua_yaml_loader *loader) {
    }
 
    if (loader->parser.context) {
-      snprintf(buf, sizeof(buf), "%s at line: %d, column: %d\n",
+      snprintf(buf, sizeof(buf), "%s at line: %zd, column: %zd\n",
          loader->parser.context,
          loader->parser.context_mark.line + 1,
          loader->parser.context_mark.column + 1);
@@ -569,7 +569,7 @@ static void dump_document(struct lua_yaml_dumper *dumper) {
    yaml_emitter_emit(&dumper->emitter, &ev);
 }
 
-static int append_output(void *arg, unsigned char *buf, unsigned int len) {
+static int append_output(void *arg, unsigned char *buf, size_t len) {
    struct lua_yaml_dumper *dumper = (struct lua_yaml_dumper *)arg;
    luaL_addlstring(&dumper->yamlbuf, (char *)buf, len);
    return 1;
